requirements.txt
Django==2.2.24
django-cors-headers==3.7.0
django-jsonfield-backport==1.0.3
djangorestframework==3.12.4
djorm-ext-pool==0.8.2
greenlet==1.1.0
importlib-metadata==4.6.1
mysqlclient==2.0.3
pytz==2021.1
SQLAlchemy==1.4.22
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.5.0
